fix avatar loading on namehash-labs LP - Our Supporters section
